got data to persist in database with JPA response entity

+++ src/main/java/culinary/example/demo/RecipeRepository.java
@@ -2,7 +2,4 @@ package culinary.example.demo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
-
-
-}
+public interface RecipeRepository extends CrudRepository<Recipe, Long> { }
